requirements.txt
asgiref==3.4.1
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.9
dj-database-url==0.5.0
Django==3.2.9
django-cleanup==5.2.0
django-cors-headers==3.10.1
django-environ==0.8.1
django-filter==21.1
django-heroku==0.3.1
# django-mapbox-location-field==1.7.3
# django-resized==0.3.11
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
guardian==0.2.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.2
Jinja2==2.11.3
Markdown==3.3.6
MarkupSafe==2.0.1
Pillow==8.4.0
psycopg2==2.9.2
psycopg2-binary==2.9.3
pycodestyle==2.8.0
PyJWT==1.7.1
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
zipp==3.6.0
